requirements.txt
alembic==1.13.3
blinker==1.9.0
click==8.1.7
colorama==0.4.6
coverage==7.6.1
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.1.2
zipp==3.20.2
